+++ src/main/java/com/spirnt/mission/discodeit/config/ScheduleConfig.java
@@ -1,0 +1,10 @@
+package com.spirnt.mission.discodeit.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@Configuration
+@EnableScheduling
+public class ScheduleConfig {
+
+}
